refactor: Consolidate all utilities into a single file
This refactoring combines all individual utility functions from the 'utils' directory into a single barrel file, 'utils/index.js'.

This simplifies the project structure by moving from a modular, multi-file approach to a single, monolithic utility file. All original utility files have been removed, and the new 'index.js' file now contains the entire utility engine with named exports for each function.
